Drop Pulp 2 plugin
* Collect context in integration test

When the plugin is not running, it is useful to have the log buffer for
the failed modules. This contains the error message while loading the
module, making it much easier to debug.

Also asserting that the response is ok is a good thing before anything
else.

* Add requires to make test standalone

This makes the following work:

    rake test TEST=test/pulpcore_plugin_integration_test.rb

* Use built in Smart Proxy URL validation

Since Smart Proxy 2.1.0 there is a built in URL validator.

* Don't use any Pulp 2 resources in PulpcoreClient

Previously it wasn't very clear where it got its config from. The
PulpcoreApi implementation used the Pulp 3 URL but printed the Pulp 2
URL in case an error showed up. Now it only uses the Pulpcore plugin.

a5501f96aa181582d0b7fd37c58283d5fd91d677 dropped disk space usage but
the require was forgotten.

* Drop Pulp 2 code

* Use rackup_path DSL method

Smart Proxy 2.3.0 introduced rackup_path which sets both
http_rackup_path and https_rackup_path to the same value. Since this is
what's done here, it simplifies the plugin definition.

* Use a short hand definition for lambda

There is no reason to use multiple lines for this.

* Update README to match the current state