Expose the container registry URL
Today Katello constructs this URL, but we're in a far better position to
determine this. The installer can also explicitly set this since it
already knows it.

It also expands the sample config to further document what the settings
are and how to use them. While not used in production by us (the
installer overwrites it), it is helpful to provide context.