mark *_url settings as required and remove their default values
Those URLs are used by Katello and clients to connect to Pulp and
`localhost` is almost never the right choice here as it will be either
not reachable (for clients) or have the wrong certs (for Katello).

Guessing the right URL is hard, so let's not do that and force the user
to set it instead.